subscriptions: update resource with db data

+++ http/subscriptions_api.go
@@ -4,6 +4,7 @@ import "net/http"
 import "github.com/rafaelbandeira3/uniqush-push/rest"
 import "github.com/rafaelbandeira3/uniqush-push/mysql"
 import (
+	"encoding/json"
 	"fmt"
 	"github.com/gorilla/mux"
 	"strconv"
@@ -14,8 +15,10 @@ type SubscriptionsApi struct {
 }
 
 func (api *SubscriptionsApi) AddDeliveryPointToService(w http.ResponseWriter, r *http.Request) {
-	resource := new(rest.SubscriptionResource)
-	readJson(r, resource)
+	resource := MustGetSubscriptionResource(w, r)
+	if resource == nil {
+		return
+	}
 
 	service, err := api.db.FindServiceByAlias(resource.ServiceAlias)
 	if err != nil {
@@ -33,14 +36,22 @@ func (api *SubscriptionsApi) AddDeliveryPointToService(w http.ResponseWriter, r 
 		return
 	}
 
-	resource.Id = id
+	err = api.UpdateResourceState(id, resource)
+	if err != nil {
+		w.WriteHeader(500)
+		jsonError := rest.JsonError{Error: "Updated but can't select subscription", GoError: err.Error()}
+		respondJson(w, jsonError)
+		return
+	}
 
 	respondJson(w, resource)
 }
 
 func (api *SubscriptionsApi) RemoveDeliveryPointFromService(w http.ResponseWriter, r *http.Request) {
-	resource := new(rest.SubscriptionResource)
-	readJson(r, resource)
+	resource := MustGetSubscriptionResource(w, r)
+	if resource == nil {
+		return
+	}
 
 	err := api.db.DeleteSubscriptionByDeviceKey(resource.Alias, resource.DeviceKey)
 	if err != nil {
@@ -54,9 +65,7 @@ func (api *SubscriptionsApi) RemoveDeliveryPointFromService(w http.ResponseWrite
 }
 
 func (api *SubscriptionsApi) UpdateDeliveryPoint(w http.ResponseWriter, r *http.Request) {
-	resource := new(rest.SubscriptionResource)
-	readJson(r, resource)
-
+	resource := MustGetSubscriptionResource(w, r)
 	vars := mux.Vars(r)
 
 	id, _ := strconv.Atoi(vars["id"])
@@ -68,7 +77,7 @@ func (api *SubscriptionsApi) UpdateDeliveryPoint(w http.ResponseWriter, r *http.
 		return
 	}
 
-	subs, err := api.db.FindSubscription(int64(id))
+	err = api.UpdateResourceState(int64(id), resource)
 	if err != nil {
 		w.WriteHeader(500)
 		jsonError := rest.JsonError{Error: "Updated but can't select subscription", GoError: err.Error()}
@@ -76,16 +85,61 @@ func (api *SubscriptionsApi) UpdateDeliveryPoint(w http.ResponseWriter, r *http.
 		return
 	}
 
-	SubscriptionFromDbToResource(subs, resource)
-
 	respondJson(w, resource)
 }
 
-func SubscriptionFromDbToResource(subs *mysql.Subscription, resource *rest.SubscriptionResource) {
+func (api *SubscriptionsApi) UpdateResourceState(id int64, resource *rest.SubscriptionResource) error {
+	subs, err := api.db.FindSubscription(id)
+	if err != nil {
+		return err
+	}
 	resource.Id = subs.Id
 	resource.Alias = subs.Alias
 	resource.PushServiceProviderType = subs.PushServiceProviderType
 	resource.DeviceKey = subs.DeviceKey
 	resource.Enabled = subs.Enabled
 	resource.ServiceAlias = subs.Service.Alias
+	return err
+}
+
+func GetSubscriptionResource(r *http.Request) (*rest.SubscriptionResource, error) {
+	resource := new(rest.SubscriptionResource)
+	var parsed map[string]interface{}
+	readJson(r, &parsed)
+
+	fmt.Println(parsed)
+
+	enabled, ok := parsed["enabled"].(bool)
+	if !ok {
+		enabledInt, ok := parsed["enabled"].(float64)
+
+		if !ok {
+			return resource, fmt.Errorf("Invalid value for 'enabled'")
+		}
+		enabled = enabledInt != 0
+	}
+	parsed["enabled"] = enabled
+
+	raw, err := json.Marshal(parsed)
+	if err != nil {
+		return resource, err
+	}
+
+	err = json.Unmarshal(raw, resource)
+	if err != nil {
+		return resource, err
+	}
+
+	return resource, nil
+}
+
+func MustGetSubscriptionResource(w http.ResponseWriter, r *http.Request) *rest.SubscriptionResource {
+	resource, err := GetSubscriptionResource(r)
+	if err != nil {
+		w.WriteHeader(422)
+		jsonError := rest.JsonError{Error: "Can't parse subscription", GoError: err.Error()}
+		respondJson(w, jsonError)
+		return nil
+	}
+	return resource
 }
